fix(api): keep Effect types out of the public serve-app declarations so emitted root d.ts stays consumer-resolvable
